refactor: add list flags
